Split and decompile some `sat` `SND` functions
