+++ config/saturn/zero.bin.yaml
@@ -75,9 +75,23 @@ segments:
       - start: 0xD9EC
         type: c
         file: zero
-        end: 0x17233
+        end: 0x14647
+
+      - start: 0x14648
+        type: c
+        file: lib/snd
+        end: 0x14FF3
+
+      - start: 0x14FF4
+        type: c
+        file: zero_2
+        end: 0x17213
 
       # DGFS
+      - start: 0x17214
+        type: data
+        file: lib/gfs
+        end: 0x17233
       - start: 0x17234
         type: c
         file: lib/gfs
