+++ config/saturn/zero_user_syms.txt
@@ -146,6 +146,9 @@ _d_060505E8 = 0x060505E8;
 _d_060576AC = 0x060576AC;
 _d_060578A0 = 0x060578A0;
 _DAT_06057A10 = 0x06057A10;
+_adr_com_block = 0x06057C68;
+_adr_host_int_work = 0x06057C6C;
+_intrflag = 0x06057C70;
 _mstCmd = 0x06057D7C;
 _slvCmd = 0x06057D80;
 _VRAM = 0x06057D84;
