+++ src/saturn/lib/snd.h
@@ -1,0 +1,23 @@
+#ifndef SND_H
+#define SND_H
+
+#include "xpt.h"
+
+#define SND_RET_SET 0
+#define SND_RET_NSET 1
+
+typedef Uint8 SndAreaMap;
+typedef Uint8 SndTlVl;
+typedef Uint8 SndEfctBnkNum;
+typedef Uint8 SndToneBnkNum;
+typedef Uint8 SndMixBnkNum;
+typedef Uint8 SndEfctOut;
+typedef Uint8 SndLev;
+typedef Sint8 SndPan;
+typedef Uint8 SndRet;
+typedef Uint8 SndHardPrm;
+typedef Uint16 SndHardStat;
+
+typedef Uint8 SndPcmNum;
+
+#endif
